+++ src/api/shared/client.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { tokenStorage } from './tokenStorage'
+import { tokenStorage } from '../login/tokenStorage'
 
 const client = axios.create({
   baseURL: import.meta.env.VITE_API_BASE_URL,
